// MyMusic.Client/src/model/song.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * MyMusic.Server | v1
 * OpenAPI spec version: 1.0.0
 */

import type { Album } from "./album";
import type { Artist } from "./artist";
import type { Genre } from "./genre";

export interface Song {
	id: number;
	/** @nullable */
	cover: number | null;
	title: string;
	artists: Artist[];
	album: Album;
	genres: Genre[];
	/** @nullable */
	year: number | null;
	duration: string;
}